Standard refactoring of some core Python types so that their implementation (including subclasses) do not access the underlying backing field, but uses an accessor. More fixes for #1645.
In addition, there's support for defaultdict providing atomic
construction of values for associated keys, using Guava MapMaker's
makeComputingMap. This is a rather cool example of the additional
abstraction provided by the refactoring (namely
PyDictionary#getMap). Some tests are needed to demo how sweet this
really is :)

Note that we cannot, given our constructors calling super in a chain,
readily remove extraneous fields that are not used, either through the
use of abstract classes to share implementation; or violate safe
publication by calling it later, outside the constructor. (There are
in fact tests that fail without safe publication.) But this is a minor
concern.