Do not emit duplicate entries in building standalone jar.
Such duplicates in Jython's case are seen in META-INF (given that it's
a global namespace, unlike in packages themselves) and can be readily
be excluded as it is now done in build.xml with this fix, using
excludes="META-INF/**".

Some of the regrtest can also now be run in such built jars (and is
now included). In particular, `java -jar jython.jar -m test.regrtest`
will not work for standalone because the directory of scripts cannot
be listed, given it is in a jar; see related issues
http://bugs.jython.org/issue1879 and http://bugs.jython.org/issue2143

Fixes http://bugs.jython.org/issue2196