Support noinherit directive for generating derived classes.
Adding constructors is not possible with the rest directive, since
this is inheritable, as can be seen with defaultdict.derived
inheriting from dict.derived. So this change adds a similar directive,
noinherit, which otherwise acts like the rest directive. Additionally,
the import directive no longer inherits, given that it's also used for
fixups of the generated derived class.