Fixes paging for the help function on Windows
On Windows, the text pager for the help function was not being
properly selected, due to os.name/os._name differences. However, even
if this OS testing is fixed, the use of subprocess to use the more
command to page was hanging.

Now uses a JLine2 specific pager if JLine2 is available that emulates
a subset of less functionality, namely paging backwards, otherwise a
plainpager (simply write lines to stdout).

Fixes http://bugs.jython.org/issue2209