Generalize adding special slot descriptors (__dict__, __weakref__)
Jython used to be too strict, and would throw TypeError if the special
slot descriptors of __dict__ and __weakref__ were used in a
subclass. With this change, CPython is stricter than Jython (see tests
in test_slots_jy for specifics). However Jython's underlying object
model is based on Java, and that give us more flexibility. In
particular, it's now possible to add a __dict__ slot descriptor to a
subclass of Java classes, such as java.util.HashMap.

Fixes http://bugs.jython.org/issue2272 and adds support for Werkzeug