Fix stack overflow due to how proxies for Java classes resolve super method in Java
For a Python class extending a Java class and/or interfaces, the MRO
contains a proxy (PyJavaType) so that Java can call back into
Python. However, because the proxy is in the MRO, it participates in
super lookup, and this was breaking because during super calls, the
referred Python code would refer to this proxy, which would refer back
to Python, generating a stack overflow.  So break out of this infinite
loop by ignoring this entry for super purposes.  The use of the
super__ prefix parallels the workaround seen in PyReflectedFunction.

Fixes http://bugs.jython.org/issue1540