Fix various threading issues introduced by previous fix
repr() of threading.{Condition,Lock,RLock} is now a cleaned-up version
of what is seen in CPython. Example for an acquired RLock:
<_threading.RLock owner='MainThread' count=1>. But we differ from
CPython in providing something similar for Lock: <_threading.Lock
owner='MainThread' locked=True>

Fixed typo such that creating a new RLock in Python would actually
return a Lock, which has quite different semantics in that it's not
actually reentrant.

Awaiting conditions may return early in general, although spurious
notifies are rare in the usual happy path. However, to avoid flakiness
in the test, do not expect the time to be exact. This requires
patching Lib/lock_tests.py specifically for this scenario.