Many compatibility fixes for launcher (bin/jython, bin/jython.exe)
Now supports CLASSPATH, and correctly parses JAVA_OPTS and JYTHON_OPTS
for inclusion into the command line for org.python.util.jython, which
is the actual main program for command line Jython.

Updated installer so it generates a shebang line for the launcher that
is specific to the CPython 2.7 installation that's available. This
change allows for -E to be passed as an option, which does not work if
shebang uses #!/usr/bin/env to select, as it previously did.

More robust testing, including unit tests that was ported from the now
removed test_bat_jy, but now applied to all platforms.

RC2 introduced a bug where jython -classpath CLASSPATH could work;
this has been removed in favor of jython -J-classpath CLASSPATH (old
behavior); this interferes with the processing of the -c option.

Fixes http://bugs.jython.org/issue2311 and http://bugs.jython.org/issue2319